rename entities to schemas
