+++ src/CSharp/EasyMicroservices.Cores.Database/Database/Schemas/FullAbilityIdSchema.cs
@@ -1,0 +1,16 @@
+﻿using EasyMicroservices.Cores.Interfaces;
+
+namespace EasyMicroservices.Cores.Database.Schemas
+{
+    /// <summary>
+    /// 
+    /// </summary>
+    /// <typeparam name="TId"></typeparam>
+    public class FullAbilityIdSchema<TId> : FullAbilitySchema, IIdSchema<TId>
+    {
+        /// <summary>
+        /// 
+        /// </summary>
+        public TId Id { get; set; }
+    }
+}
