+++ src/CSharp/EasyMicroservices.Cores.Database/Database/Schemas/LongIdSchema.cs
@@ -1,0 +1,12 @@
+﻿using EasyMicroservices.Cores.Interfaces;
+
+namespace EasyMicroservices.Cores.Database.Schemas
+{
+    /// <summary>
+    /// 
+    /// </summary>
+    public class LongIdSchema : IdSchema<long>
+    {
+
+    }
+}
