+++ src/CSharp/EasyMicroservices.Cores.Database/EasyMicroservices.Cores.Database.csproj
@@ -24,6 +24,7 @@
   </ItemGroup>
 
   <ItemGroup>
+    <Folder Include="Database\Entities\" />
     <Folder Include="Database\ReadableLogics\" />
     <Folder Include="Database\WritableLogics\" />
   </ItemGroup>
